refactor(core): move classes to new modules
